requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
image==1.5.33
joblib==1.2.0
lightgbm==2.2.3
mysqlclient==2.1.1
numpy==1.23.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
pandas==1.5.1
Pillow==9.2.0
pybase64==1.2.3
python-dateutil==2.8.2
pytz==2022.4
scikit-learn==1.1.2
scipy==1.9.2
six==1.16.0
sklearn==0.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2022.4
xgboost==0.90
